Introduce cell type for TodayGrid
This is not generic and doesn't really allow a custom view to be used
and so is not really usable outside of this instance. But for the time
being may be okay.

I think a better design may be one like:
```swift
TodayGrid(within: width, contents: dataObjects, layoutPattern: pattern) { (object, sizeKind) in
  // determine the view to return based on the object
  // sizeKind would be an instance of the objects in pattern but with the true
  // final dimensions available so that fine grained layout can be done
}
```